rename admin-login to admin-login-router
