+++ api/server.js
@@ -6,7 +6,7 @@ const usersRouter = require("../users/user-router");
 const contentRouter = require("../author-content/content-router");
 const libraryRouter = require("../content-library/library-router");
 const adminRegisterRouter = require("../auth/admin-register-router");
-const adminLoginRouter = require("../auth/admin-login");
+const adminLoginRouter = require("../auth/admin-login-router");
 const adminRouter = require("../admins/admin-router");
 const userResetPassword = require("../users/user-reset-password");
 const messageRouter = require("../messaging/message-router");
